+++ packages/TreeViewOld/src/index.vue
@@ -1,0 +1,167 @@
+<template>
+    <div :class="['fv-' + $theme + '-TreeView']" :style="{ background }" ref="view">
+        <tree-content :style="style" :children="items" :deepth="0" :viewStyle="style" :checkable="checkable"
+            :padding="space" :draggable="draggable" :foreground="foreground" :borderWidth="borderWidth"
+            :revealEffect="revealEffect" :expandedIcon="expandedIcon" :unexpandedIcon="unexpandedIcon"
+            :expandedIconPosition="expandedIconPosition" :background="background" @click="click"
+            :backgroundColorHover="backgroundColorHover" :backgroundColorActive="backgroundColorActive"
+            :expandClickMode="expandClickMode">
+            <template v-slot:default="prop">
+                <slot :item="prop.item"> </slot>
+            </template>
+        </tree-content>
+    </div>
+</template>
+
+<script>
+import '../../office-ui-fabric-core/dist/css/fabric.min.css';
+import TreeContent from './components/content';
+
+export default {
+    name: 'FvTreeView',
+    components: {
+        TreeContent,
+    },
+    props: {
+        theme: {
+            type: String,
+            default: 'system',
+        },
+        checkable: {
+            type: Boolean,
+            default: false,
+        },
+        data: {
+            default: function () {
+                return [];
+            },
+            required: true,
+            type: Array,
+        },
+        space: {
+            default: 20,
+            type: Number,
+        },
+        borderWidth: {
+            default: 2,
+        },
+        revealEffect: {
+            type: Boolean,
+            default: true,
+        },
+        draggable: {
+            type: Boolean,
+            default: false,
+        },
+        viewStyle: {},
+        expandedIcon: {
+            type: String,
+            default: 'ChevronDownMed',
+        },
+        unexpandedIcon: {
+            type: String,
+            default: 'ChevronUpMed',
+        },
+        foreground: {
+            type: String,
+        },
+        background: {
+            type: String,
+            default: undefined,
+        },
+        expandedIconPosition: {
+            type: String,
+            default: 'left',
+        },
+        backgroundColorHover: {
+            type: String,
+        },
+        backgroundColorActive: {
+            type: String
+        },
+        expandClickMode: {
+            type: String,
+            default: "normal"
+        },
+        itemHeight: {
+            type: String,
+        }
+    },
+    data() {
+        return {
+            style: {},
+            changeLock: false,
+        };
+    },
+    model: {
+        prop: 'data',
+        event: 'update:data',
+    },
+    computed: {
+        $theme() {
+            if (this.theme === 'system') return this.$fvGlobal.state.theme;
+            return this.theme;
+        },
+        items: {
+            get: function () {
+                return this.data;
+            },
+            set: function (val) {
+                this.$emit('update:data', val);
+            },
+        },
+    },
+    watch: {
+        $theme() {
+            this.$nextTick(() => {
+                this.initStyle();
+            });
+        },
+        data: {
+            handler() {
+                if (this.changeLock) return;
+                this.changeLock = true;
+                this.$emit('change', this.data);
+                setTimeout(() => {
+                    this.changeLock = false;
+                }, 100);
+            },
+            deep: true,
+        },
+        viewStyle: {
+            handler() {
+                this.initStyle();
+            },
+            deep: true,
+        },
+    },
+    mounted() {
+        this.initStyle();
+    },
+    methods: {
+        initStyle() {
+            if (document.defaultView) {
+                this.style = this.viewStyle || {
+                    backgroundColor: document.defaultView.getComputedStyle(this.$refs.view, null).backgroundColor,
+                    color: document.defaultView.getComputedStyle(this.$refs.view, null).color,
+                };
+            }
+        },
+        mutexSelected(item, items) {
+            for (let index in items) {
+                let el = items[index];
+                if (el != item && el.selected == true) {
+                    el.selected = false;
+                }
+                if (el.children) {
+                    this.mutexSelected(item, el.children);
+                }
+            }
+        },
+        click(item) {
+            if (!this.checkable) this.mutexSelected(item, this.items);
+            this.$emit('click', item);
+        },
+    },
+};
+</script>
