+++ packages/common-theme/TreeView.scss
@@ -1,68 +1,47 @@
 @mixin fv-TreeView {
   position: relative;
-  width: 100%;
-  min-width: 200px;
-  list-style: none;
-  margin: 0;
-  padding: 0px 3px;
-  user-select: none;
-  box-sizing: border-box;
-  .fv-TreeView__item {
-    cursor: default;
-    width: 100%;
-    margin: 0;
-    padding: 0;
-    color: inherit;
-    background-color: inherit;
-    .fv-TreeView__label-border {
-      padding: 0px;
-      height: 30px;
-      margin: 4px 0;
-      .fv-TreeView__label {
-        --fv-TreeView__label-hover-color: initial;
-        --fv-TreeView__label-color: initial;
-        position: relative;
-        border-radius: 5px;
-        color: inherit;
-        padding: 0 0px;
-        border: transparent solid 2px;
-        box-sizing: border-box;
-        height: 30px;
-        display: inline-flex;
-        align-items: center;
-        width: 100%;
-        i {
-          padding: 0;
-          margin: 0 6px 0 0;
-          font-size: 12px;
-        }
-        .fv-TreeView__checkbox {
-          margin: 0 6px 0 0;
-        }
-        .fv-TreeView__icon {
-          font-size: 16px;
-          width: auto;
-          margin: 0 12px;
-        }
-        img.fv-TreeView__icon {
-          height: 24px;
-        }
-        .fv-TreeView__text {
-          display: inline-block;
-          user-select: none;
-        }
-        .fv-TreeView__right-box {
-          position: absolute;
-          right: 10px;
-        }
-      }
-    }
-  }
-  .fv-TreeView__tree {
+  padding: 0;
+  ul,
+  li {
     list-style: none;
     padding: 0;
     margin: 0;
-    background-color: inherit;
-    color: inherit;
+  }
+  .fv-TreeView--item {
+    .fv-TreeView--item-field {
+      --fv-TreeView--item-field-padding-left: 0px;
+      padding-left: var(--fv-TreeView--item-field-padding-left);
+      display: flex;
+      align-items: center;
+      .fv-TreeView--item-checkbox {
+        margin-right: 5px;
+      }
+      .fv-TreeView--item-expand {
+        width: 16px;
+        margin-right: 10px;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        cursor: pointer;
+        font-size: 10px;
+      }
+      .fv-TreeView--item-label {
+        width: 100%;
+        display: flex;
+        align-items: center;
+        .fv-TreeView--item-label-icon {
+          height: 20px;
+          width: 20px;
+          padding: 0px;
+          margin: 0 4px;
+          font-size: 16px;
+          display: flex;
+          align-items: center;
+          &img{
+            object-fit: cover;
+          }
+        }
+      }
+    }
   }
 }
