+++ packages/dark-theme/TreeView.scss
@@ -1,31 +1,4 @@
 @import "../common-theme/TreeView.scss";
 .fv-dark-TreeView {
   @include fv-TreeView();
-  color: #fff;
-  background-color: #000;
-  .fv-TreeView__label {
-    &:hover {
-      background-color: var(--fv-TreeView__label-hover-color, rgba(255, 255, 255, 0.1));
-    }
-    &.active {
-      background-color: var(--fv-TreeView__label-color, rgba(255, 255, 255, 0.1));
-    }
-  }
-  .fv-TreeView__checkbox {
-    border-color: #666;
-    &:hover {
-      border-color: #d6d6d6;
-    }
-    &.checked {
-      background-color: #2f2f2f;
-      border-color: #2f2f2f;
-    }
-    &.Indeterminate {
-      background-color: #2f2f2f;
-      border-color: #2f2f2f;
-    }
-    i {
-      color: white;
-    }
-  }
 }
